Major changes

+++ frontend/crop_recommendation_system/lib/Chatbot/chatbot_service.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'package:crop_recommendation_system/Authentication/secure_storage_service.dart';
 import 'package:http/http.dart' as http;
 import 'conversation_model.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
@@ -9,41 +10,62 @@ class ApiConfig {
 
 class ChatbotService {
   Future<Map<String, dynamic>> sendMessage({
-    required String userId,
     required String domain,
-    required String language,
     required String query,
     String? conversationId,
   }) async {
+    final token = await SecureStorageService.getAccessToken();
+
+    print("Token: $token");
+
+    if (token == null) {
+      throw Exception("User is not logged in");
+    }
+
     final response = await http.post(
       Uri.parse("${ApiConfig.baseUrl}/chat"),
-      headers: {"Content-Type": "application/json"},
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": "Bearer $token",
+      },
       body: jsonEncode({
-        "user_id": userId,
         "domain": domain,
-        "language": language,
         "query": query,
         "conversation_id": conversationId,
       }),
     );
 
+    print("Status Code: ${response.statusCode}");
+    print("Response Body: ${response.body}");
+
     if (response.statusCode == 200) {
       return jsonDecode(response.body);
     }
 
-    throw Exception("Failed to send message");
+    throw Exception(
+      "Failed to send message\n"
+      "Status: ${response.statusCode}\n"
+      "Body: ${response.body}",
+    );
   }
 
   Future<List<ConversationModel>> getConversations({
-    required String userId,
     required String domain,
   }) async {
+    final token = await SecureStorageService.getAccessToken();
+
+    print("Token: $token");
+
+    if (token == null) {
+      throw Exception("User is not logged in");
+    }
+
     final response = await http.get(
       Uri.parse(
         "${ApiConfig.baseUrl}/conversations"
-        "?user_id=$userId"
-        "&domain=$domain",
+        "?domain=$domain",
       ),
+      headers: {"Authorization": "Bearer $token"},
     );
 
     if (response.statusCode == 200) {
@@ -52,12 +74,25 @@ class ChatbotService {
       return (data as List).map((e) => ConversationModel.fromJson(e)).toList();
     }
 
-    throw Exception("Failed to load conversations");
+    throw Exception(
+      "Failed to load conversations"
+      "Status: ${response.statusCode}\n"
+      "Body: ${response.body}",
+    );
   }
 
   Future<dynamic> getConversation(String conversationId) async {
+    final token = await SecureStorageService.getAccessToken();
+
+    print("Token: $token");
+
+    if (token == null) {
+      throw Exception("User is not logged in");
+    }
+
     final response = await http.get(
       Uri.parse("${ApiConfig.baseUrl}/conversations/$conversationId"),
+      headers: {"Authorization": "Bearer $token"},
     );
 
     print(response.body);
@@ -73,10 +108,21 @@ class ChatbotService {
     required String conversationId,
     required String title,
   }) async {
+    final token = await SecureStorageService.getAccessToken();
+
+    print("Token: $token");
+
+    if (token == null) {
+      throw Exception("User is not logged in");
+    }
+
     final response = await http.patch(
       Uri.parse("${ApiConfig.baseUrl}/conversations/$conversationId"),
 
-      headers: {"Content-Type": "application/json"},
+      headers: {
+        "Content-Type": "application/json",
+        "Authorization": "Bearer $token",
+      },
 
       body: jsonEncode({"title": title}),
     );
@@ -90,8 +136,17 @@ class ChatbotService {
   }
 
   Future<void> deleteConversation({required String conversationId}) async {
+    final token = await SecureStorageService.getAccessToken();
+
+    print("Token: $token");
+
+    if (token == null) {
+      throw Exception("User is not logged in");
+    }
+
     final response = await http.delete(
       Uri.parse("${ApiConfig.baseUrl}/conversations/$conversationId"),
+      headers: {"Authorization": "Bearer $token"},
     );
 
     if (response.statusCode != 200) {
